grpc/requirements.txt
grpcio-tools
google-gax
google-auth
google-auth-httplib2
requests